removed console.log, updated comments
