+++ client/src/pages/private/SiteTable.js
@@ -183,7 +183,6 @@ export default ({ sites, viewOnly }) => {
   };
 
   const columnObj = (rowId) => columns.find(({ id }) => id === rowId);
-  console.log(rows);
   return (
     <>
       <NewSiteDialog
