+++ client/src/services/index.js
@@ -4,3 +4,4 @@ export * from './return-of-service';
 export * from './feature-flags';
 export * from './notifications';
 export * from './participant-details';
+export * from './allocations';
